Make Dock, ShipGenerator, Hobos and other

+++ lab-02/by/Dzenia/docks_and_hobos/Persons/Ship.java
@@ -2,9 +2,9 @@ package by.Dzenia.docks_and_hobos.Persons;
 
 public class Ship {
     private final Cargo cargo;
-    private final long weight;
+    private final int weight;
 
-    public Ship(String cargoType, long weight) {
+    public Ship(String cargoType, int weight) {
         cargo = new Cargo(cargoType);
         this.weight = weight;
     }
@@ -13,7 +13,7 @@ public class Ship {
         return cargo;
     }
 
-    public long getWeight() {
+    public int getWeight() {
         return weight;
     }
 }
